+++ ui/src/alphabets.test.ts
@@ -4,6 +4,8 @@ import { detectAlphabet, translate } from './alphabets';
 test('simple alphabet test', ({ expect }) => {
   expect(detectAlphabet(['ATTAGACA'])?.type).toEqual('nucleotide');
   expect(detectAlphabet(['attagaca'])?.type).toEqual('nucleotide');
+  expect(detectAlphabet(['attagacawrr'])?.type).toEqual('nucleotide');
+  expect(detectAlphabet(['atnagacawrr'])?.type).toEqual('nucleotide');
   expect(detectAlphabet(['CASSLAPGAT'])?.type).toEqual('amino-acid');
   expect(detectAlphabet(['casslapgat'])?.type).toEqual('amino-acid');
   expect(detectAlphabet(['hello world'])?.type).toBeUndefined();